requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
# python-dotenv
numpy
opencv-python==4.7.0.68
Pillow==9.1.1
requests==2.28.1
